# shared/swagger.json
{"openapi":"3.0.0","paths":{"/authentication/login":{"post":{"operationId":"AuthenticationController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/authentication/register":{"post":{"operationId":"AuthenticationController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/authentication/status":{"get":{"operationId":"AuthenticationController_status","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Authentication"]}},"/api/residence":{"post":{"operationId":"ResidenceController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResidenceDto"}}}},"responses":{"201":{"description":""}},"tags":["Residence"]},"get":{"operationId":"ResidenceController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Residence"]}},"/api/residence/{id}":{"get":{"operationId":"ResidenceController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Residence"]},"patch":{"operationId":"ResidenceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResidenceDto"}}}},"responses":{"200":{"description":""}},"tags":["Residence"]},"delete":{"operationId":"ResidenceController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Residence"]}},"/api/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/incident":{"post":{"operationId":"IncidentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIncidentDto"}}}},"responses":{"201":{"description":""}},"tags":["Incident"]},"get":{"operationId":"IncidentController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Incident"]}},"/api/incident/{id}":{"get":{"operationId":"IncidentController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Incident"]},"patch":{"operationId":"IncidentController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateIncidentDto"}}}},"responses":{"200":{"description":""}},"tags":["Incident"]},"delete":{"operationId":"IncidentController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Incident"]}},"/api/vote-session":{"post":{"operationId":"VoteSessionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVoteSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["VoteSession"]},"get":{"operationId":"VoteSessionController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["VoteSession"]}},"/api/vote-session/{id}":{"get":{"operationId":"VoteSessionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["VoteSession"]},"patch":{"operationId":"VoteSessionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVoteSessionDto"}}}},"responses":{"200":{"description":""}},"tags":["VoteSession"]},"delete":{"operationId":"VoteSessionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["VoteSession"]}},"/api/document":{"post":{"operationId":"DocumentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDocumentDto"}}}},"responses":{"201":{"description":""}},"tags":["Document"]},"get":{"operationId":"DocumentController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Document"]}},"/api/document/{id}":{"get":{"operationId":"DocumentController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Document"]},"patch":{"operationId":"DocumentController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDocumentDto"}}}},"responses":{"200":{"description":""}},"tags":["Document"]},"delete":{"operationId":"DocumentController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Document"]}},"/api/vote":{"post":{"operationId":"VoteController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVoteDto"}}}},"responses":{"201":{"description":""}},"tags":["Vote"]},"get":{"operationId":"VoteController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Vote"]}},"/api/vote/{id}":{"get":{"operationId":"VoteController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vote"]},"patch":{"operationId":"VoteController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVoteDto"}}}},"responses":{"200":{"description":""}},"tags":["Vote"]},"delete":{"operationId":"VoteController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vote"]}},"/api/announcement":{"post":{"operationId":"AnnouncementController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAnnouncementDto"}}}},"responses":{"201":{"description":""}},"tags":["Announcement"]},"get":{"operationId":"AnnouncementController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Announcement"]}},"/api/announcement/{id}":{"get":{"operationId":"AnnouncementController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Announcement"]},"patch":{"operationId":"AnnouncementController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAnnouncementDto"}}}},"responses":{"200":{"description":""}},"tags":["Announcement"]},"delete":{"operationId":"AnnouncementController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Announcement"]}}},"info":{"title":"Blockwise API","description":"API for copropriété management","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"residenceId":{"type":"string"},"role":{"type":"string","enum":["admin","resident","board","guest"],"default":"resident"}},"required":["email","password","firstName","lastName","phoneNumber","residenceId","role"]},"CreateResidenceDto":{"type":"object","properties":{}},"UpdateResidenceDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"residenceId":{"type":"string"},"role":{"type":"string","enum":["admin","resident","board","guest"],"default":"resident"}},"required":["email","password","firstName","lastName","phoneNumber","residenceId","role"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"residenceId":{"type":"string"},"role":{"type":"string","enum":["admin","resident","board","guest"],"default":"resident"}}},"CreateIncidentDto":{"type":"object","properties":{}},"UpdateIncidentDto":{"type":"object","properties":{}},"CreateVoteSessionDto":{"type":"object","properties":{}},"UpdateVoteSessionDto":{"type":"object","properties":{}},"CreateDocumentDto":{"type":"object","properties":{}},"UpdateDocumentDto":{"type":"object","properties":{}},"CreateVoteDto":{"type":"object","properties":{}},"UpdateVoteDto":{"type":"object","properties":{}},"CreateAnnouncementDto":{"type":"object","properties":{}},"UpdateAnnouncementDto":{"type":"object","properties":{}}}}}